stylished the spotify player
