+++ client/src/pages/Player.tsx
@@ -1,14 +1,11 @@
-import SpotifyPlayer from "../components/SpotifyPlayer";
-import { useFetchDataContext } from "../context/FetchDataContext";
 import styles from "./Player.module.css";
 import { useState } from "react";
 import DisplayPopularityFilterOnPlayer from "../components/DisplayPopularityFilterOnPlayer";
 import DisplaySelectedTagsOnPlayer from "../components/DisplaySelectedTagsOnPlayer";
 import ArtistInfo from "../components/ArtistInfo";
+import TrackPlayingCard from "../components/TrackPlayingCard";
 
 const Player = () => {
-  const { musicList } = useFetchDataContext();
-
   const [isOpen, setIsOpen] = useState(false);
   const toggleButton = () => {
     setIsOpen(!isOpen);
@@ -16,25 +13,9 @@ const Player = () => {
 
   return (
     <>
-      <section className={styles.GlobalPlayer}>
-        {musicList.length > 0 ? (
-          musicList.map((item) => (
-            <div key={item.id}>
-              <p className={styles.playerId}>ID: {item.id}</p>
-              <div className={styles.windowPlayer}>
-                <SpotifyPlayer uri={item.url} />
-                <a
-                  href={item.url}
-                  target="_blank"
-                  rel="noopener noreferrer"
-                ></a>
-              </div>
-            </div>
-          ))
-        ) : (
-          <p>No data available</p>
-        )}
-      </section>
+      <div>
+        <TrackPlayingCard />
+      </div>
       <div>
         <DisplaySelectedTagsOnPlayer />
       </div>
